refactor(all): improve code organization
Split classes out into separete files and sort exported classes above private helpers.
